Initial transfer of Common to Geometry
