+++ Common_Engine/Common_Engine.csproj
@@ -71,20 +71,7 @@
     <Compile Include="Modify\SetInternalElements2D.cs" />
     <Compile Include="Modify\SetOutlineElements1D.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Query\Area.cs" />
-    <Compile Include="Query\Bounds.cs" />
-    <Compile Include="Query\Centroid.cs" />
-    <Compile Include="Query\ControlPoints.cs" />
-    <Compile Include="Query\ElementCurves.cs" />
-    <Compile Include="Query\ElementVertices.cs" />
-    <Compile Include="Query\Elements0D.cs" />
-    <Compile Include="Query\Geometry.cs" />
-    <Compile Include="Query\InternalElements2D.cs" />
-    <Compile Include="Query\IsSelfIntersecting.cs" />
-    <Compile Include="Query\Length.cs" />
     <Compile Include="Query\Normal.cs" />
-    <Compile Include="Query\OutlineElements1D.cs" />
-    <Compile Include="Query\OutlineCurve.cs" />
     <Compile Include="Query\ShearModulus.cs" />
   </ItemGroup>
   <ItemGroup>
