+++ Geometry_Engine/Query/ControlPoints.cs
@@ -92,6 +92,34 @@ namespace BH.Engine.Geometry
             return curve.ControlPoints;
         }
 
+        /******************************************/
+        /****            IElement1D            ****/
+        /******************************************/
+
+        public static List<Point> ControlPoints(this IElement1D element1D)
+        {
+            return IControlPoints(element1D.IGeometry());
+        }
+
+
+        /******************************************/
+        /****            IElement2D            ****/
+        /******************************************/
+
+        public static List<Point> ControlPoints(this IElement2D element2D, bool externalOnly = false)
+        {
+            List<Point> pts = ControlPoints(element2D.IOutlineCurve());
+            if (!externalOnly)
+            {
+                foreach (IElement2D e in element2D.IInternalElements2D())
+                {
+                    pts.AddRange(e.ControlPoints());
+                }
+            }
+            return pts;
+        }
+
+        /******************************************/
 
         /***************************************************/
         /**** Public Methods - Interfaces               ****/
