+++ Geometry_Engine/Query/OutlineElements1D.cs
@@ -20,12 +20,10 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Common;
 using BH.oM.Geometry;
 using System.Collections.Generic;
-using System.Linq;
 
-namespace BH.Engine.Common
+namespace BH.Engine.Geometry
 {
     public static partial class Query
     {
@@ -33,23 +31,9 @@ namespace BH.Engine.Common
         /****            IElement2D            ****/
         /******************************************/
 
-        public static PolyCurve IOutlineCurve(this IElement2D element2D)
+        public static List<IElement1D> IOutlineElements1D(this IElement2D element2D)
         {
-            return new PolyCurve { Curves = element2D.IOutlineElements1D().Select(e => e.IGeometry()).ToList() };
-        }
-
-        /******************************************/
-
-        public static PolyCurve IOutlineCurve(this List<IElement1D> elements1D)
-        {
-            return new PolyCurve { Curves = elements1D.Select(e => e.IGeometry()).ToList() };
-        }
-
-        /******************************************/
-
-        public static List<PolyCurve> IInternalOutlineCurves(this IElement2D element2D)
-        {
-            return element2D.IInternalElements2D().Select(x => x.IOutlineCurve()).ToList();
+            return Reflection.Compute.RunExtensionMethod(element2D, "OutlineElements1D") as List<IElement1D>;
         }
 
         /******************************************/
